+++ container.go
@@ -91,8 +91,8 @@ type ContainerRequest struct {
 	ReaperImage     string              // alternative reaper image
 	AutoRemove      bool                // if set to true, the container will be removed from the host when stopped
 	NetworkMode     container.NetworkMode
-	AlwaysPullImage bool               // Always pull image
-	BuildArgs       map[string]*string // for specifying network aliases
+	AlwaysPullImage bool // Always pull image
+	BuildArgs       map[string]*string
 }
 
 // ProviderType is an enum for the possible providers
